requirements.txt
streamlit>=1.32
numpy
opencv-python-headless
pillow
tensorflow
torch
ultralytics
matplotlib
pandas
scikit-learn
requests
